multiple: prerequisite struct changes for dm-verity for snaps (#16168)
* s/integrity: export dmverity hash file name helper

* snap/info.go: add IntegrityData field to Info struct

Although a snap can have multiple integrity data available from the
store, only one is active on the current system so this is not
a list.

* o/snapstate: add IntegrityData to SnapSetup

* snap: add new DmVerityInfo method for ContainerPlaceInfo interface

DmVerityInfo is used to construct the dm-verity data file path from
the snap's mount file and the root hash in its integrity data.

* multiple: rename SnapSetup.IntegrityData to IntegrityDataParams

There are now several fields named IntegrityData used in various
contexts. To help with readability, renaming the IntegrityData field
inside the SnapSetup struct, the context that is passed around during
a snap's installation, to IntegrityDataParams.

* store: add IntegrityData field in storeSnap

a snap from the store can have multiple integrity data available in the
future. Currently only one type of integrity data is served of type
"dm-verity" using "sha256" so the first element of the list is selected
when creating the snap Info struct.

* snap: rename DmVerityInfo to DmVerityParamsIfPresent

* snap: rename IntegrityData struct to IntegrityDataInfo

* snap: name return values of DmVerityParamsIfPresent

* o/snapstate/backend_test.go: add channel with integrity data

* o/snapstate: rename SnapSetup.IntegrityDataParams to SnapSetup.IntegrityDataInfo

* snap: split DmVerityParams API

* multiple: embed integrity.IntegrityDataParams in snap.IntegrityDataInfo

* o/snapstate: change json serialization of IntegrityDataInfo

* snap: change json serialization of IntegrityDataInfo params field

* store/details_v2.go: use ParseUint instead of Atoi

* store/details_v2.go: clarify usage of uint size for dm-verity parameters

* store: add integrity field in the list of excluded fields for find

* store: add integrity field in the list of excluded fields for info

* store: add integrity field in the list of excluded fields for the older details API