cmd/snap: rename version output fields indicating source of client and daemon binaries
Rename the fields which indicate where the snap and the snapd binaries
come from. This should make the output more accessible, even if we keep
asking people only for 'snap version' output and the version they have
may not yet support --verbose flag. However, keep the verbose flag
around in case anyone finds the PR that introduced it.

Signed-off-by: Maciej Borzecki <maciej.borzecki@canonical.com>