cmd/snap-confine: handle potential race when creating /tmp/snap-private-tmp
When /tmp/snap-private-tmp does not exist, it will be created at runtime
by snap-confine. Specifically, when more than one instance of
snap-confine is active, both race to create the directory. Commit
https://github.com/canonical/snapd/commit/ba1b2ff7a73e0c2a41d0b9ca3dc78863ad4a78e4
changed the code to use sc_ensure_mkdir() helper. The helper, creates
the directory with 0000 mode bits, and proceeds to update its ownership
and mode to match what was requested. It is possible that when one
instance of snap-confine is in the middle of creating the directory,
another instance may observe an intermediate state in which either the
ownership or permissions have not yet been updated. Instead of failing,
let the other instance attempt to amend the ownership and permissions.

Related: SNAPDENG-36092
Fixes: LP#2134364

Signed-off-by: Maciej Borzecki <maciej.borzecki@canonical.com>