core-initrd: add systemd service for setting plymouth keyboard layout (#16269)
* Add service for setting plymouth keymap

If `snapd.xkb` is set on the kernel command line, this service
will use the set value to create `/etc/default/keyboard` before
Plymouth starts. This will allow Plymouth to consider layout changes
between boots.

The service's script is written to be POSIX compliant with the
BusyBox shell.

* Add plymouth keymap service symlink in sysinit

This should make the keymap service start alongside plymouth
and remove the need for an [Install] section.

* Fix plymouth-set-keymap ordering and comment

Allow Plymouth to start before executing `plymouth-set-keymap.service`
and add comment explaining why `/etc/default/keyboard` exists.

* Fix bash string and revert Before service

This will make the static checks happier with respect to `xkb_list`.
and enforce that plymouth is enabled.
Also revert Before to plymouth-start.service for clarity.

* Fix xkb_list splitting and multiple options

A previous commit added double quotes to xkb_list, however
this had the side effect of not splitting the variable properly.
This fix uses the read call instead to avoid the use of double
qoutes. Additionally, this reads everything after options as options
with separators.

* Create /etc/default in set-plymouth-keymap

By default, core-initrd does not create /etc/default, but the
keyboard file depends on the existence of this directory.