secboot: use device node to find the storage container and keys
We do not want to use a symlink as we do not want to rely on udev having
created them at this point in the boot process. Use directly the device node
that we already now. We still need to use paths for the symlinks when calling
WithLegacyKeyringKeyDescriptionPaths for the descriptions in the kernel
keyring, but they are not accessed because of this.