tests: move machines in ps6 to ps7 (#16363)
* Move machines in ps6 to ps7

Also include support for opensuse 16

* remove all the sufix -ps7 from the backends

* remove support for os16

* remove consolidated-report file

* remove more opensuse 16.0 bits

* remove packaging file for opensuse 16

* pushing missing changes removing os-16

* Update .github/workflows/ci-spread-times.yaml

Co-authored-by: Katie May <katie.m.may@protonmail.com>

* avoid running the full run-spread nested test when it is updated

The test will fail but it won't run the full set of tests. Also it has
been removed the full variant as it is not used and it is triggering a
full run as well.

* Extend the selinux policy to allow connecting to squid

We already allow connecting to generic HTTP proxies (classified as
http_cache_port_t). Extend the policy to allow connecting to squid
proxies (squid_port_t), in case someone is using them (like we do in
tests running in PS7).

---------

Co-authored-by: Katie May <katie.m.may@protonmail.com>